refactor: répartition de DataService.ts
