+++ src/features/Scolarite/screens/WebBrowserScreen.tsx
@@ -8,7 +8,7 @@ import { GestureDetector, Gesture } from 'react-native-gesture-handler';
 
 import style, { tokens } from '../../../shared/theme/Theme';
 import { AppContext } from '../../../shared/services/AppCore';
-import { URL } from '../../../shared/services/DataService';
+import { URL } from '../../../shared/constants/urls';
 import SecureStoreService from '../../../shared/services/SecureStoreService';
 import Translator from '../../../shared/i18n/Translator';
