+++ src/features/Settings/screens/AboutScreen.tsx
@@ -7,7 +7,7 @@ import { MaterialCommunityIcons } from '@expo/vector-icons';
 import style, { tokens } from '../../../shared/theme/Theme';
 import Translator from '../../../shared/i18n/Translator';
 import { AppContext } from '../../../shared/services/AppCore';
-import { URL } from '../../../shared/services/DataService';
+import { URL } from '../../../shared/constants/urls';
 import { withHeaderAnimation } from '../../../shared/navigation/NavHelpers';
 
 export interface SectionProps {
