+++ src/shared/constants/urls.ts
@@ -1,0 +1,18 @@
+export const URL = {
+    MAP: 'https://www.google.com/maps/',
+    CONTACT_EMAIL: 'mailto:contact@kaelab.dev',
+    UKIT_WEBSITE: 'https://ukit-bordeaux.fr',
+    KAELAB_WEBSITE: 'https://kaelab.dev',
+    LEGAL_NOTICE: 'https://github.com/KAE-Lab/UKit/blob/master/PRIVACY.md',
+    VERSION_STORE: 'https://raw.githubusercontent.com/KAE-Lab/UKit/master/VERSION',
+    GOOGLE_APP: 'https://play.google.com/store/apps/details?id=com.bordeaux1.emplois',
+    APPLE_APP: 'https://apps.apple.com/fr/app/ukit-bordeaux/id1394708917',
+    CROUSTILLANT_WEBSITE: 'https://croustillant.menu',
+};
+
+export const WebApiURL = {
+    DOMAIN: 'https://ukit.kbdev.io/Home/',
+    GROUPS: 'ReadResourceListItems',
+    CALENDARDATA: 'GetCalendarData',
+    SIDEBAR: 'GetSideBarEvent',
+};
